+++ apps/manifest-demo/3010-rspack-provider/src/components/ButtonOldAnt.tsx
@@ -1,9 +1,10 @@
 import Button from 'antd/lib/button';
 import { version } from 'antd/package.json';
-import stuff from './stuff.module.css';
+import * as stuff from './stuff.module.css';
 
 export default function ButtonOldAnt() {
   return (
+    // @ts-ignore
     <Button className={stuff['test-remote2'] + ' test-remote2'}>
       Button from antd@{version}
     </Button>