+++ packages/enhanced/src/lib/container/ContainerPlugin.ts
@@ -8,11 +8,11 @@ import ContainerEntryModuleFactory from './ContainerEntryModuleFactory';
 import ContainerExposedDependency from './ContainerExposedDependency';
 import { parseOptions } from './options';
 import type {
-  optimize,
   Compiler,
   Compilation,
   WebpackError,
   WebpackPluginInstance,
+  WebpackPluginFunction,
 } from 'webpack';
 import type { containerPlugin } from '@module-federation/sdk';
 import FederationRuntimePlugin from './runtime/FederationRuntimePlugin';
@@ -20,20 +20,12 @@ import FederationModulesPlugin from './runtime/FederationModulesPlugin';
 import checkOptions from '../../schemas/container/ContainerPlugin.check';
 import schema from '../../schemas/container/ContainerPlugin';
 import FederationRuntimeDependency from './runtime/FederationRuntimeDependency';
+import type { OptimizationSplitChunksCacheGroup } from 'webpack/lib/optimize/SplitChunksPlugin';
+import type { Falsy } from 'webpack/declarations/WebpackOptions';
 
 const ModuleDependency = require(
   normalizeWebpackPath('webpack/lib/dependencies/ModuleDependency'),
 ) as typeof import('webpack/lib/dependencies/ModuleDependency');
-
-type ExcludeUndefined<T> = T extends undefined ? never : T;
-type NonUndefined<T> = ExcludeUndefined<T>;
-
-type OptimizationSplitChunksOptions = NonUndefined<
-  ConstructorParameters<typeof optimize.SplitChunksPlugin>[0]
->;
-
-type CacheGroups = OptimizationSplitChunksOptions['cacheGroups'];
-type CacheGroup = NonUndefined<CacheGroups>[string];
 
 const createSchemaValidation = require(
   normalizeWebpackPath('webpack/lib/util/create-schema-validation'),
@@ -77,13 +69,21 @@ class ContainerPlugin {
       ),
       runtimePlugins: options.runtimePlugins,
       experiments: options.experiments,
+      dataPrefetch: options.dataPrefetch,
     };
   }
 
   // container should not be affected by splitChunks
   static patchChunkSplit(compiler: Compiler, name: string): void {
     const { splitChunks } = compiler.options.optimization;
-    const patchChunkSplit = (cacheGroup: CacheGroup) => {
+    const patchChunkSplit = (
+      cacheGroup:
+        | string
+        | false
+        | ((...args: any[]) => any)
+        | RegExp
+        | OptimizationSplitChunksCacheGroup,
+    ) => {
       switch (typeof cacheGroup) {
         case 'boolean':
         case 'string':
@@ -160,7 +160,7 @@ class ContainerPlugin {
 
   apply(compiler: Compiler): void {
     const useModuleFederationPlugin = compiler.options.plugins.find(
-      (p: WebpackPluginInstance) => {
+      (p: Falsy | WebpackPluginInstance | WebpackPluginFunction) => {
         if (typeof p !== 'object' || !p) {
           return false;
         }
@@ -190,13 +190,15 @@ class ContainerPlugin {
 
     compiler.hooks.make.tapAsync(
       PLUGIN_NAME,
-      (
+      async (
         compilation: Compilation,
         callback: (error?: WebpackError | null | undefined) => void,
       ) => {
         const hasSingleRuntimeChunk =
           compilation.options?.optimization?.runtimeChunk;
         const hooks = FederationModulesPlugin.getCompilationHooks(compilation);
+        const federationRuntimeDependency =
+          federationRuntimePluginInstance.getDependency(compiler);
         const dep = new ContainerEntryDependency(
           name,
           //@ts-ignore
@@ -204,23 +206,46 @@ class ContainerPlugin {
           shareScope,
           federationRuntimePluginInstance.entryFilePath,
           this._options.experiments,
+          this._options.dataPrefetch,
         );
         dep.loc = { name };
-        compilation.addEntry(
-          compilation.options.context || '',
-          dep,
-          {
-            name,
-            filename,
-            runtime: hasSingleRuntimeChunk ? false : runtime,
-            library,
-          },
-          (error: WebpackError | null | undefined) => {
-            if (error) return callback(error);
-            hooks.addContainerEntryModule.call(dep);
-            callback();
-          },
-        );
+
+        await new Promise((resolve, reject) => {
+          compilation.addEntry(
+            compilation.options.context || '',
+            dep,
+            {
+              name,
+              filename,
+              runtime: hasSingleRuntimeChunk ? false : runtime,
+              library,
+            },
+            (error: WebpackError | null | undefined) => {
+              if (error) return reject(error);
+              hooks.addContainerEntryModule.call(dep);
+              resolve(undefined);
+            },
+          );
+        }).catch(callback);
+
+        await new Promise((resolve, reject) => {
+          compilation.addInclude(
+            compiler.context,
+            federationRuntimeDependency,
+            { name: undefined },
+            (err, module) => {
+              if (err) {
+                return reject(err);
+              }
+              hooks.addFederationRuntimeModule.call(
+                federationRuntimeDependency,
+              );
+              resolve(undefined);
+            },
+          );
+        }).catch(callback);
+
+        callback();
       },
     );
 
@@ -228,56 +253,53 @@ class ContainerPlugin {
     // we have to use finishMake in order to check the entries created and see if there are multiple runtime chunks
     compiler.hooks.finishMake.tapAsync(
       PLUGIN_NAME,
-      async (compilation, callback) => {
-        // its a child compiler
+      (compilation: Compilation, callback) => {
         if (
           compilation.compiler.parentCompilation &&
           compilation.compiler.parentCompilation !== compilation
         ) {
-          // dont include dependencies on child compilations
           return callback();
         }
 
         const hooks = FederationModulesPlugin.getCompilationHooks(compilation);
-        const createdRuntimes = new Set();
+        const createdRuntimes = new Set<string>();
+
         for (const entry of compilation.entries.values()) {
-          if (entry.options.runtime) {
-            if (createdRuntimes.has(entry.options.runtime)) {
-              continue;
-            }
-            createdRuntimes.add(entry.options.runtime);
+          const runtime = entry.options.runtime;
+          if (runtime) {
+            createdRuntimes.add(runtime);
           }
         }
 
-        // if it has multiple runtime chunks - make another with no name or runtime assigned
         if (
-          createdRuntimes.size !== 0 ||
-          compilation.options?.optimization?.runtimeChunk
+          createdRuntimes.size === 0 &&
+          !compilation.options?.optimization?.runtimeChunk
         ) {
-          const dep = new ContainerEntryDependency(
-            name,
-            //@ts-ignore
-            exposes,
-            shareScope,
-            federationRuntimePluginInstance.entryFilePath,
-            this._options.experiments,
-          );
-
-          dep.loc = { name };
-
-          compilation.addInclude(
-            compilation.options.context || '',
-            dep,
-            { name: undefined },
-            (error: WebpackError | null | undefined) => {
-              if (error) return callback(error);
-              hooks.addContainerEntryModule.call(dep);
-              callback();
-            },
-          );
-        } else {
-          callback();
+          return callback();
         }
+
+        const dep = new ContainerEntryDependency(
+          name,
+          //@ts-ignore
+          exposes,
+          shareScope,
+          federationRuntimePluginInstance.entryFilePath,
+          this._options.experiments,
+          this._options.dataPrefetch,
+        );
+
+        dep.loc = { name };
+
+        compilation.addInclude(
+          compilation.options.context || '',
+          dep,
+          { name: undefined },
+          (error: WebpackError | null | undefined) => {
+            if (error) return callback(error);
+            hooks.addContainerEntryModule.call(dep);
+            callback();
+          },
+        );
       },
     );
 
@@ -301,11 +323,6 @@ class ContainerPlugin {
     compiler.hooks.thisCompilation.tap(
       PLUGIN_NAME,
       (compilation: Compilation, { normalModuleFactory }) => {
-        const federationRuntimeDependency =
-          federationRuntimePluginInstance.getDependency(compiler);
-
-        const logger = compilation.getLogger('ContainerPlugin');
-        const hooks = FederationModulesPlugin.getCompilationHooks(compilation);
         compilation.dependencyFactories.set(
           FederationRuntimeDependency,
           normalModuleFactory,
@@ -313,21 +330,6 @@ class ContainerPlugin {
         compilation.dependencyTemplates.set(
           FederationRuntimeDependency,
           new ModuleDependency.Template(),
-        );
-
-        compilation.addInclude(
-          compiler.context,
-          federationRuntimeDependency,
-          { name: undefined },
-          (err, module) => {
-            if (err) {
-              return logger.error(
-                'Error adding federation runtime module:',
-                err,
-              );
-            }
-            hooks.addFederationRuntimeModule.call(federationRuntimeDependency);
-          },
         );
       },
     );