+++ packages/dts-plugin/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @module-federation/dts-plugin
 
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies [d5a3072]
+  - @module-federation/sdk@0.6.11
+  - @module-federation/managers@0.6.11
+  - @module-federation/third-party-dts-extractor@0.6.11
+
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies [22a3b83]
+  - @module-federation/sdk@0.6.10
+  - @module-federation/managers@0.6.10
+  - @module-federation/third-party-dts-extractor@0.6.10
+
+## 0.6.9
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.9
+- @module-federation/managers@0.6.9
+- @module-federation/third-party-dts-extractor@0.6.9
+
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies [32db0ac]
+  - @module-federation/sdk@0.6.8
+  - @module-federation/managers@0.6.8
+  - @module-federation/third-party-dts-extractor@0.6.8
+
+## 0.6.7
+
+### Patch Changes
+
+- 0216364: fix(dts-plugin): ignore .vue ext
+- Updated dependencies [9e32644]
+  - @module-federation/sdk@0.6.7
+  - @module-federation/managers@0.6.7
+  - @module-federation/third-party-dts-extractor@0.6.7
+
 ## 0.6.6
 
 ### Patch Changes