+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @module-federation/esbuild
 
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [d5a3072]
+  - @module-federation/sdk@0.6.11
+
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [22a3b83]
+  - @module-federation/sdk@0.6.10
+
+## 0.0.27
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.9
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [32db0ac]
+  - @module-federation/sdk@0.6.8
+
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [9e32644]
+  - @module-federation/sdk@0.6.7
+
 ## 0.0.24
 
 ### Patch Changes