+++ packages/manifest/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @module-federation/manifest
 
+## 0.6.11
+
+### Patch Changes
+
+- 83c8620: fix(manifest): record all remotes which are declared in build config
+- Updated dependencies [d5a3072]
+  - @module-federation/sdk@0.6.11
+  - @module-federation/dts-plugin@0.6.11
+  - @module-federation/managers@0.6.11
+
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies [22a3b83]
+  - @module-federation/sdk@0.6.10
+  - @module-federation/dts-plugin@0.6.10
+  - @module-federation/managers@0.6.10
+
+## 0.6.9
+
+### Patch Changes
+
+- @module-federation/sdk@0.6.9
+- @module-federation/managers@0.6.9
+- @module-federation/dts-plugin@0.6.9
+
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies [32db0ac]
+  - @module-federation/sdk@0.6.8
+  - @module-federation/dts-plugin@0.6.8
+  - @module-federation/managers@0.6.8
+
+## 0.6.7
+
+### Patch Changes
+
+- Updated dependencies [9e32644]
+- Updated dependencies [0216364]
+  - @module-federation/sdk@0.6.7
+  - @module-federation/dts-plugin@0.6.7
+  - @module-federation/managers@0.6.7
+
 ## 0.6.6
 
 ### Patch Changes