+++ packages/modernjs/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @module-federation/modern-js
 
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies [d5a3072]
+  - @module-federation/sdk@0.6.11
+  - @module-federation/node@2.5.21
+  - @module-federation/enhanced@0.6.11
+
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies [6b02145]
+- Updated dependencies [22a3b83]
+  - @module-federation/enhanced@0.6.10
+  - @module-federation/sdk@0.6.10
+  - @module-federation/node@2.5.20
+
+## 0.6.9
+
+### Patch Changes
+
+- Updated dependencies [70a1708]
+  - @module-federation/enhanced@0.6.9
+  - @module-federation/node@2.5.19
+  - @module-federation/sdk@0.6.9
+
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies [32db0ac]
+  - @module-federation/sdk@0.6.8
+  - @module-federation/enhanced@0.6.8
+  - @module-federation/node@2.5.18
+
+## 0.6.7
+
+### Patch Changes
+
+- Updated dependencies [1b6bf0e]
+- Updated dependencies [9e32644]
+- Updated dependencies [9e32644]
+- Updated dependencies [9e32644]
+- Updated dependencies [9e32644]
+  - @module-federation/enhanced@0.6.7
+  - @module-federation/sdk@0.6.7
+  - @module-federation/node@2.5.17
+
 ## 0.6.6
 
 ### Patch Changes