+++ packages/runtime-tools/CHANGELOG.md
@@ -1,5 +1,47 @@
 ## [1.0.1-canary.1](https://github.com/module-federation/core/compare/runtime-1.0.0...runtime-1.0.1-canary.1) (2023-12-06)
 
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies [ea6d417]
+  - @module-federation/runtime@0.6.11
+  - @module-federation/webpack-bundler-runtime@0.6.11
+
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies [b704f30]
+  - @module-federation/runtime@0.6.10
+  - @module-federation/webpack-bundler-runtime@0.6.10
+
+## 0.6.9
+
+### Patch Changes
+
+- @module-federation/runtime@0.6.9
+- @module-federation/webpack-bundler-runtime@0.6.9
+
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies [32db0ac]
+- Updated dependencies [6c5f444]
+- Updated dependencies [fac6ecf]
+  - @module-federation/runtime@0.6.8
+  - @module-federation/webpack-bundler-runtime@0.6.8
+
+## 0.6.7
+
+### Patch Changes
+
+- Updated dependencies [9e32644]
+- Updated dependencies [9e32644]
+  - @module-federation/runtime@0.6.7
+  - @module-federation/webpack-bundler-runtime@0.6.7
+
 ## 0.6.6
 
 ### Patch Changes