+++ packages/webpack-bundler-runtime/CHANGELOG.md
@@ -1,5 +1,51 @@
 # [1.0.0-canary.3](https://github.com/module-federation/core/compare/webpack-bundler-runtime-1.0.0-canary.2...webpack-bundler-runtime-1.0.0-canary.3) (2023-11-23)
 
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies [ea6d417]
+- Updated dependencies [d5a3072]
+  - @module-federation/runtime@0.6.11
+  - @module-federation/sdk@0.6.11
+
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies [b704f30]
+- Updated dependencies [22a3b83]
+  - @module-federation/runtime@0.6.10
+  - @module-federation/sdk@0.6.10
+
+## 0.6.9
+
+### Patch Changes
+
+- @module-federation/runtime@0.6.9
+- @module-federation/sdk@0.6.9
+
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies [32db0ac]
+- Updated dependencies [32db0ac]
+- Updated dependencies [6c5f444]
+- Updated dependencies [fac6ecf]
+  - @module-federation/sdk@0.6.8
+  - @module-federation/runtime@0.6.8
+
+## 0.6.7
+
+### Patch Changes
+
+- Updated dependencies [9e32644]
+- Updated dependencies [9e32644]
+- Updated dependencies [9e32644]
+  - @module-federation/runtime@0.6.7
+  - @module-federation/sdk@0.6.7
+
 ## 0.6.6
 
 ### Patch Changes